+++ CentralizedSalesSystem.API/Models/User.cs
@@ -5,8 +5,8 @@ namespace CentralizedSalesSystem.API.Models
 {
     public class User
     {
-        public int Id { get; set; }
-        public int BusinessId { get; set; }
+        public long Id { get; set; }
+        public long BusinessId { get; set; }
 
         public string Email { get; set; } = string.Empty;
         public string PasswordHash { get; set; } = string.Empty;
